servers/webscraper/mcp_project/pyproject.toml
[project]
name = "webscraper-mcp"
version = "0.1.0"
description = "MCP Server for intelligent web scraping with Firecrawl and BeautifulSoup engines"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "firecrawl-py>=1.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.uv]
python-downloads = "manual"
